Stabilized unit tests
